Moved UI tests to feature dedicated mock module
