--- app-mock/build.gradle
@@ -1,44 +1,0 @@
-plugins {
-    id 'com.android.application'
-    id 'kotlin-android'
-    id 'kotlin-android-extensions'
-    id 'kotlin-kapt'
-}
-
-android {
-    compileSdkVersion 29
-
-    defaultConfig {
-        applicationId "pl.sulej.githubusers.mock"
-        minSdkVersion 17
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-    sourceSets {
-        main.res.srcDirs += '../app-res'
-    }
-}
-
-dependencies {
-    implementation kotlinStandardLibrary
-
-    implementation project(":users")
-    implementation project(":users-mock")
-
-    implementation daggerAndroid
-    kapt daggerAndroidProcessor
-    kapt daggerCompiler
-
-    androidTestImplementation espresso
-}
