+++ users-mock/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'com.android.library'
+    id 'com.android.application'
     id 'kotlin-android'
     id 'kotlin-android-extensions'
     id 'kotlin-kapt'
@@ -15,6 +15,8 @@ android {
         versionName "1.0"
 
         consumerProguardFiles "consumer-rules.pro"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
@@ -30,6 +32,9 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+    sourceSets {
+        main.res.srcDirs += '../app-res'
+    }
 }
 
 dependencies {
@@ -43,4 +48,6 @@ dependencies {
     implementation daggerAndroid
     kapt daggerCompiler
     kapt daggerAndroidProcessor
+    
+    androidTestImplementation espresso
 }
