+++ users-mock/src/main/AndroidManifest.xml
@@ -1,1 +1,22 @@
-<manifest package="pl.sulej.users.mock" />
+<manifest xmlns:tools="http://schemas.android.com/tools" package="pl.sulej.users.mock"
+          xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <uses-permission android:name="android.permission.INTERNET"/>
+
+    <application
+            android:allowBackup="true"
+            android:name="pl.sulej.users.MockUsersApplication"
+            android:icon="@mipmap/ic_launcher"
+            android:label="@string/app_name"
+            android:roundIcon="@mipmap/ic_launcher_round"
+            android:supportsRtl="true"
+            android:theme="@style/AppTheme" tools:ignore="GoogleAppIndexingWarning">
+        <activity android:name="pl.sulej.users.view.UsersActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN"/>
+
+                <category android:name="android.intent.category.LAUNCHER"/>
+            </intent-filter>
+        </activity>
+    </application>
+</manifest>
